test/unit/modules/storage/commands/update_file_command.spec.ts
import { UpdateFileCommand } from '../../../../../src/app/modules/storage/commands/update_file_command';

describe('UpdateFileCommand', () => {
  let command: UpdateFileCommand;

	beforeEach(() => {
		command = new UpdateFileCommand();
	});

	describe('constructor', () => {
		it('should have valid name', () => {
			expect(command.name).toEqual('updateFile');
		});

		it('should have valid schema', () => {
			expect(command.schema).toMatchSnapshot();
		});
	});

	describe('verify', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('execute', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
